refactor: move legacy/sandbox/ → sandbox/, legacy/core/tools.ts → agent/tools.ts
Pure file relocation with import path updates. No behavior change.
